DEMOS-613: combined amendment and extensions tables, added extra test for uncovered code
